Remove unnecessary comments and add defensive checks

+++ user_profile/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.db import IntegrityError
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, HttpResponseBadRequest
 from django.core.paginator import Paginator
 from recipe.models import Recipe, Like, Comment, Favourite
 from recipe.forms import CommentForm
@@ -13,6 +13,10 @@ from .forms import ProfileForm
 @login_required
 def dashboard(request, filter_type=None):
     user = request.user
+
+    ALLOWED_FILTERS = {"my", "favourites", None}
+    if filter_type not in ALLOWED_FILTERS:
+        return HttpResponseBadRequest("Invalid filter type.")
 
     # Recipes
     if filter_type == "my":
@@ -35,14 +39,18 @@ def dashboard(request, filter_type=None):
             user=request.user, recipe=recipe).exists()
 
     # Profile
-    profile = Profile.objects.get(user=user)
+    try:
+        profile = Profile.objects.get(user=user)
+    except Profile.DoesNotExist:
+        profile = Profile.objects.create(user=user)
+        messages.info(request, "New profile was created for your account.")
 
     if request.method == 'POST':
         profile_form = ProfileForm(request.POST, request.FILES, instance=profile)
         if profile_form.is_valid():
             profile_form.save()
             messages.success(request, "Profile updated successfully.")
-            return redirect('dashboard') 
+            return redirect('dashboard')
     else:
         profile_form = ProfileForm(instance=profile)
 
@@ -66,7 +74,7 @@ def recipe_detail(request, pk):
     recipe = get_object_or_404(queryset, id=pk)
 
     comments = recipe.comments.all().order_by("-updated_on")
-    comment_count = recipe.comments.all().count()
+    comment_count = comments.count()
 
     if request.method == "POST":
         comment_form = CommentForm(request.POST)
@@ -75,9 +83,7 @@ def recipe_detail(request, pk):
             comment.author = request.user
             comment.recipe = recipe
             comment.save()
-            messages.add_message(
-                request, messages.SUCCESS, "Comment successfully submitted"
-            )
+            messages.success(request, "Comment successfully submitted.")
 
     comment_form = CommentForm()
 
@@ -134,63 +140,71 @@ def recipe_unlike(request, pk):
 
 @login_required
 def comment_edit(request, recipe_id, comment_id):
-    if request.method == "POST":
+    if request.method != "POST":
+        messages.warning(request, "Invalid request method.")
+        return redirect("recipe_detail", recipe_id)
 
+    if request.method == "POST":
         recipe = get_object_or_404(Recipe, id=recipe_id)
         comment = get_object_or_404(Comment, id=comment_id)
         comment_form = CommentForm(request.POST, instance=comment)
+
+        if comment.author != request.user:
+            messages.error(request, "You are not allowed to edit this comment.")
+            return redirect("recipe_detail", recipe_id)
 
         if comment_form.is_valid() and comment.author == request.user:
             comment = comment_form.save(commit=False)
             comment.recipe = recipe
             comment.author = request.user
             comment.save()
-            messages.add_message(
-                request, messages.SUCCESS, 'Comment successfully updated.')
+            messages.success(request, "Comment successfully updated.")
         else:
-            messages.add_message(
-                request, messages.ERROR,
-                'Error updating comment. Please try again.')
+            messages.error(request, "Error updating comment. Please try again.")
 
     return HttpResponseRedirect(reverse('recipe_detail', args=[recipe_id]))
 
 
 @login_required
 def comment_delete(request, recipe_id, comment_id):
+    if request.method != "POST":
+        messages.warning(request, "Invalid request method.")
+        return redirect("recipe_detail", recipe_id)
+
     comment = get_object_or_404(Comment, id=comment_id)
 
     if comment.author == request.user:
         comment.delete()
-        messages.add_message(
-            request, messages.SUCCESS, 'Comment successfully deleted!')
+        messages.success(request, "Comment successfully deleted!")
     else:
-        messages.add_message(
-            request, messages.ERROR, 'You can only delete your own comments!')
+        messages.error(request, "You can only delete your own comments!")
 
     return HttpResponseRedirect(reverse('recipe_detail', args=[recipe_id]))
 
 
 @login_required
 def add_to_favourites(request, recipe_id):
+    if request.method != "POST":
+        messages.warning(request, "Invalid request method.")
+        return redirect("dashboard")
+
     recipe = get_object_or_404(Recipe, id=recipe_id)
 
     if Favourite.objects.filter(user=request.user, recipe=recipe).exists():
-        messages.add_message(
-                request, messages.INFO,
-                "This recipe is already in your favourites."
-            )
+        messages.info(request, "This recipe is already in your favourites.")
     else:
         Favourite.objects.create(user=request.user, recipe=recipe)
-        messages.add_message(
-                request, messages.SUCCESS,
-                "Recipe successfully added to favourites."
-            )
+        messages.success(request, "Recipe successfully added to favourites.")
 
     return redirect("dashboard")
 
 
 @login_required
 def remove_from_favourites(request, recipe_id):
+    if request.method != "POST":
+        messages.warning(request, "Invalid request method.")
+        return redirect("dashboard")
+
     recipe = get_object_or_404(Recipe, id=recipe_id)
     favourite = Favourite.objects.filter(
         user=request.user, recipe=recipe).first()
